cleaned up files and tweaked output logs
